OMAP34XX: CAM: Remove the need for RESUME state
In RESUME power state, we were only switching a GPIO for that, and we
weren't executing again the whole POWER_ON commands.

Sakari Ailus removed the RESUME power state from his patchset,
rendering the RESUME uncompilable. So, for compensating that, i'm
storing now the previous power state, to know what is the intended
transition:

- Off -> On, or
- standby -> on (which is the resume case now)

This patch Keeps the same functionality without the need for a resume
state.

Signed-off-by: Sergio Aguirre <saaguirre@ti.com>