ARM: OMAP3: Automatic TV detection
This patch provides automatic TV detection support
on omapzoom display driver. When a stream is run on
video1, video2 is automatically linked and displayed on TV.

User can enable or disable this bind to TV, through the
use of /sys/class/display_control/omap_disp_control/tv_state,
automatic detection/switching is only enabled when this sysfs
entry is populated with 1 and a TV is connected to S-Video port.

Signed-off-by: Axel Castaneda Gonzalez<x0055901@ti.com>