From: Jagadeesh Bhaskar Pakaravoor <j-pakaravoor@ti.com
This adds the core interrupt functionality for TWL6030. A new config
CONFIG_TWL6030_CORE is added along with new files twl6030_core.c and
twl6030_irq.c.

TWL6030 still exports functions in the name of twl4030_*
and uses MACROS/functions/variables of the form twl4030_*. This needs
to be fixed.

Signed-off-by: Jagadeesh Bhaskar Pakaravoor <j-pakaravoor@ti.com>